pyproject.toml
[tool.poetry]
name = "pyfish"
version = "0.1.0"
description = ""
authors = ["raph <raphuzanjr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
